rename and fix container styles
