+++ src/ignitus-Routes/ignitus-UserInterfaceBookRoutes/index.tsx
@@ -132,7 +132,7 @@ export const UserInterfaceBookRoutes: React.FunctionComponent = () => (
         <Route path="/interface/userProfile" component={interfaceUserProfile} />
         <Route path="/interface/modalHooray" component={interfaceModalHooray} />
         <Route path="/interface/modalShare" component={interfaceModalShare} />
-        <Route path="/interface/overlayLogout" component={interfaceOverlay} />
+        <Route path="/interface/overlay" component={interfaceOverlay} />
         <Route path="/interface/scrollBar" component={interfaceScrollBar} />
         <Route
           path="/interface/toggleButtons"
