+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Molecules/ignitus-Overlay/Components/index.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import * as O from '../styles';
 import { OverlayProps } from '../types';
-import { AppIcon } from '../../../../..';
+import { AppIcon } from '../../../..';
 
-export const OverlayLogout = ({ primaryText, secondaryText }: OverlayProps) => (
+export const Overlay = ({ primaryText, secondaryText }: OverlayProps) => (
   <O.Overlay>
-    <O.TopSection>
+    <O.MessageSection>
       <O.Logo name={AppIcon.IgnitusLogo} />
       <O.Paragraph>{primaryText}</O.Paragraph>
       <O.Paragraph secondary>{secondaryText}</O.Paragraph>
-    </O.TopSection>
+    </O.MessageSection>
   </O.Overlay>
 );
