+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Molecules/ignitus-Overlay/styles.ts
@@ -1,34 +1,29 @@
 import styled from '@emotion/styled';
 import { Props } from './types';
-import { flexibleColDiv } from '../../../shared';
-import { Paragraph as P } from '../../../..';
-import { Icon } from '../../../../ignitus-Utilities/Components/icon';
+import { flexibleColDiv } from '../../shared';
+import { Paragraph as P } from '../../..';
+import { Icon } from '../../../ignitus-Utilities/Components/icon';
 
 export const Overlay = styled.div`
   position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
   background-color: rgba(0, 0, 102, 0.89);
   backdrop-filter: blur(10px);
-  width: 62rem;
-  height: 35rem;
+  width: 100%;
+  height: 100%;
 `;
 
 export const Logo = styled(Icon)`
-  padding: 2rem;
-  width: 5rem;
-  height: 5rem;
+  margin-bottom: 4rem;
+  height: 7.5rem;
   path {
     fill: white;
   }
 `;
 
-export const TopSection = styled(flexibleColDiv)`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: flex;
-  padding: 2rem;
-`;
+export const MessageSection = flexibleColDiv;
 
 export const Paragraph = styled(P)<Props>`
   color: white;
