+++ src/ignitus-UserInterfaceBook/Components/Molecules/interfaceOverlay/Components/index.tsx
@@ -1,14 +1,21 @@
 import React from 'react';
-import { OverlayLogout, Heading2 } from '../../../../../ignitus-Shared';
+import { Overlay, Heading2 } from '../../../../../ignitus-Shared';
 import { Interface } from '../../../../styles';
 
 export const interfaceOverlay = () => (
   <Interface>
     <Heading2>Overlay</Heading2>
     <hr />
-    <OverlayLogout
-      primaryText="You are now logged out."
-      secondaryText="See you soon!"
-    />
+    <div
+      style={{
+        width: '60vw',
+        height: '70vh',
+      }}
+    >
+      <Overlay
+        primaryText="You are now logged out."
+        secondaryText="See you soon!"
+      />
+    </div>
   </Interface>
 );
